Fix showToast with app.js
