+++ package.json
@@ -3,7 +3,6 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@bootstrap-styled/v4": "^3.4.0",
     "@emotion/react": "^11.4.0",
     "@fortawesome/fontawesome-svg-core": "^1.2.35",
     "@fortawesome/free-solid-svg-icons": "^5.15.3",
@@ -12,19 +11,15 @@
     "@testing-library/react": "^11.2.6",
     "@testing-library/user-event": "^12.8.3",
     "axios": "^0.21.1",
-    "bootstrap-styled": "^2.7.2",
     "firebase": "^8.6.1",
-    "firebaseui": "^4.8.0",
     "highcharts": "^9.1.0",
     "highcharts-react-official": "^3.0.0",
-    "lodash": "^4.17.21",
     "prop-types": "^15.7.2",
     "rc-tooltip": "^5.1.1",
     "react": "^17.0.2",
     "react-content-loader": "^6.0.3",
     "react-custom-scrollbars-2": "^4.4.0",
     "react-dom": "^17.0.2",
-    "react-grid-layout": "^1.2.5",
     "react-input-autosize": "^3.0.0",
     "react-redux": "^7.2.4",
     "react-router": "^5.2.0",
@@ -33,7 +28,6 @@
     "react-spinners": "^0.11.0",
     "redux": "^4.1.0",
     "styled-components": "^4.4.1",
-    "styled-reset": "^4.3.4",
     "styled-system": "^5.1.5",
     "web-vitals": "^1.1.2"
   },
