+++ src/component/Header.js
@@ -1,4 +1,10 @@
-import React, { createRef, useEffect, useRef, useState } from "react";
+import React, {
+  createRef,
+  useEffect,
+  useRef,
+  useState,
+  useContext,
+} from "react";
 import styled from "styled-components";
 import { color, space, typography } from "styled-system";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
@@ -10,10 +16,8 @@ import Sign from "./Sign";
 import Forget from "./Forget";
 import { subscribeUserData, firebaseAuthSignOut } from "../Utils/firebase";
 import { updatePageName } from "../Redux/Actions/actionCreator";
-import Toast from "./Toast";
-import checkIcon from "../images/check.svg";
-import errorIcon from "../images/error.svg";
 import logo from "../images/cryptoLogo.svg";
+import Context from "../context/Context";
 
 const Navigation = styled.header`
   width: 100%;
@@ -156,66 +160,19 @@ const Navigation = styled.header`
 const Header = () => {
   const [isExpanded, setIsExpanded] = useState(false);
   const [email, setemail] = useState(null);
-  const [list, setList] = useState([]);
   const [uid, setUid] = useState(null);
   const [loginStatus, setLoginStatus] = useState("Sign In");
-  let toastProperties = null;
   const [signType, setSignType] = useState("signin");
   const history = useHistory();
   const page = useSelector((state) => state.pageReducer.name);
   const signModal = useRef(null);
   const forgetModal = createRef();
   const dispatch = useDispatch();
+  const showToast = useContext(Context);
 
   const handleToggle = (e) => {
     e.preventDefault();
     setIsExpanded(!isExpanded);
-  };
-
-  const showToast = (type) => {
-    const id = Math.floor(Math.random() * 101 + 1);
-    switch (type) {
-      case "successSignOut":
-        toastProperties = {
-          id,
-          title: "Success signout",
-          description: "Successful signout",
-          backgroundColor: "#5cb85c",
-          icon: checkIcon,
-        };
-        break;
-      case "dangerPortfolio":
-        toastProperties = {
-          id,
-          title: "Please signin",
-          description: "Before accessing the portfolio page, please signin",
-          backgroundColor: "#d9534f",
-          icon: errorIcon,
-        };
-        break;
-      case "successSignIn":
-        toastProperties = {
-          id,
-          title: "Success signin",
-          description: "Successful signin",
-          backgroundColor: "#5cb85c",
-          icon: checkIcon,
-        };
-        break;
-      case "successSignUp":
-        toastProperties = {
-          id,
-          title: "Success signup",
-          description: "Successful signup",
-          backgroundColor: "#5cb85c",
-          icon: checkIcon,
-        };
-        break;
-      default:
-        setList([]);
-    }
-
-    setList([...list, toastProperties]);
   };
 
   const handleClickSignOut = () => {
@@ -335,7 +292,6 @@ const Header = () => {
           getSignInfo={getSignInfo}
         />
       </CustomModal>
-      <Toast toastList={list} autoDelete dismissTime={3000} />
     </>
   );
 };
