+++ src/context/Context.js
@@ -1,0 +1,5 @@
+import React from "react";
+
+const Context = React.createContext();
+
+export default Context;
